Changed config embed and changed maxpeople to maxjoins and timebetweenfiles to timespan

+++ antiraider.rb
@@ -15,6 +15,13 @@ def writeconfig
     return
   end
   puts "Writing config..."
+  if File.exist?('./config.json')
+    configfile = File.read('./config.json')
+    confighash = JSON.parse(configfile)
+    confighash['token'] = token
+    File.open("./config.json", 'w') { |file| file.write(JSON.dump(confighash)) }
+    return
+  end
   jsonwrite = "{\"token\":\"#{token}\"}"
   File.open("./config.json", 'w') { |file| file.write(jsonwrite) }
 end
@@ -80,7 +87,7 @@ bot.ready do |_|
   bot.watching = "for raids"
 end
 
-bot.command :config do |event, setting, option|
+bot.command :config, description: 'Configure the bot' do |event, setting, option|
   unless event.user.permission?(:administrator)
     event.channel.send_embed do |embed|
       embed.colour = 0xFF0000
@@ -90,13 +97,35 @@ bot.command :config do |event, setting, option|
     next
   end
   unless setting
+    configfile = File.read('./config.json')
+    confighash = JSON.parse(configfile)
+    if confighash['timespan'] == nil
+      timespan = "300"
+    else
+      timespan = Integer(confighash['timespan']) * 60
+    end
+    if confighash['maxjoins'] == nil
+      maxjoins = "10"
+    else
+      maxjoins = Integer(confighash['maxjoins'])
+    end
+
     event.channel.send_embed do |embed|
       embed.colour = 0x0080FF
       embed.title = 'Settings'
-      embed.description = "timebetweenfiles: time in minutes\nmaxpeople: number"
+      embed.fields = [
+        Discordrb::Webhooks::EmbedField.new(
+          name: 'maxjoins: number',
+          value: "Currently set to: #{maxjoins}"
+        ),
+        Discordrb::Webhooks::EmbedField.new(
+          name: 'timespan: time in minutes',
+          value: "Currently set to: #{timespan}"
+        )
+      ]
     end
   end
-  if setting == 'timebetweenfiles'
+  if setting == 'timespan'
     unless option
       event.channel.send_embed do |embed|
         embed.colour = 0xFF0000
@@ -114,14 +143,14 @@ bot.command :config do |event, setting, option|
     end
     configfile = File.read('./config.json')
     confighash = JSON.parse(configfile)
-    confighash['timebetweenfiles'] = option
+    confighash['timespan'] = option
     File.open("./config.json", 'w') { |file| file.write(JSON.dump(confighash)) }
     event.channel.send_embed do |embed|
       embed.colour = 0x0080FF
-      embed.title = "Set `timebetweenfiles` to `#{option}`"
+      embed.title = "Set `timespan` to `#{option}`"
     end
     next
-  elsif setting == 'maxpeople'
+  elsif setting == 'maxjoins'
     unless option
       event.channel.send_embed do |embed|
         embed.colour = 0xFF0000
@@ -139,11 +168,11 @@ bot.command :config do |event, setting, option|
     end
     configfile = File.read('./config.json')
     confighash = JSON.parse(configfile)
-    confighash['maxpeople'] = option
+    confighash['maxjoins'] = option
     File.open("./config.json", 'w') { |file| file.write(JSON.dump(confighash)) }
     event.channel.send_embed do |embed|
       embed.colour = 0x0080FF
-      embed.title = "Set `maxpeople` to `#{option}`"
+      embed.title = "Set `maxjoins` to `#{option}`"
     end
     next
   end
@@ -152,18 +181,18 @@ end
 bot.member_join do |event|
   configfile = File.read('./config.json')
   confighash = JSON.parse(configfile)
-  timebetweenfiles = confighash['timebetweenfiles'] * 60
-  if timebetweenfiles == nil
-    timebetweenfiles = "300"
+  timespan = confighash['timespan'] * 60
+  if timespan == nil
+    timespan = "300"
   end
-  maxpeople = confighash['maxpeople']
-  if maxpeople == nil
-    maxpeople = "10"
+  maxjoins = confighash['maxjoins']
+  if maxjoins == nil
+    maxjoins = "10"
   end
   if File.exist?("temp/#{event.server.id}.log")
     filetime = Time.parse(File.foreach("temp/#{event.server.id}.log").first)
     difference = Time.now - filetime
-    if difference >= timebetweenfiles
+    if difference >= timespan
       file = File.open("temp/#{event.server.id}.log", "w")
       file.write(Time.now)
       file.write("\n#{event.user.id}")
@@ -174,7 +203,7 @@ bot.member_join do |event|
     file.write("\n#{event.user.id}")
     file.close
     linecount = File.open("temp/#{event.server.id}.log", "r").each_line.count - 1
-    if linecount >= maxpeople
+    if linecount >= maxjoins
       event.server.kick(event.user, 'Might be a raider - Detected by AntiRaider')
     end
   else
